scripts/requirements.txt
requests
pandas
pyyaml
google-cloud-bigquery
functions-framework
pytest